Updated type for all I2C slave addresses

+++ libraries/abstractions/common_io/test/iot_test_common_io_internal.h
@@ -232,12 +232,12 @@
 
 /* I2C */
 #if defined( IOT_TEST_COMMON_IO_I2C_SUPPORTED ) && ( IOT_TEST_COMMON_IO_I2C_SUPPORTED >= 1 )
-    extern uint8_t uctestIotI2CSlaveAddr;         /* Address of Slave I2C (7-bit address) connected to the bus */
-    extern uint8_t uctestIotI2CDeviceRegister;    /* Slave I2C register address used in read/write tests */
-    extern uint8_t uctestIotI2CWriteVal;          /* Write value that will be used in the register write test */
-    extern uint8_t uctestIotI2CInstanceIdx;       /* I2C instance used in the test */
-    extern uint8_t uctestIotI2CInstanceNum;       /* The total number of I2C instances on the device */
-    extern uint8_t ucAssistedTestIotI2CSlaveAddr; /* The slave address to be set for the assisted test. */
+    extern uint16_t uctestIotI2CSlaveAddr;         /* Address of Slave I2C (7-bit address) connected to the bus */
+    extern uint8_t uctestIotI2CDeviceRegister;     /* Slave I2C register address used in read/write tests */
+    extern uint8_t uctestIotI2CWriteVal;           /* Write value that will be used in the register write test */
+    extern uint8_t uctestIotI2CInstanceIdx;        /* I2C instance used in the test */
+    extern uint8_t uctestIotI2CInstanceNum;        /* The total number of I2C instances on the device */
+    extern uint16_t ucAssistedTestIotI2CSlaveAddr; /* The slave address to be set for the assisted test. */
 
 /**
  * Board specific I2C config set
